distributed_q_learning/requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attr==0.3.2
attrs==23.2.0
boto3==1.34.103
botocore==1.34.103
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
crowdai_api==0.1.22
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.1
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.14.0
flatbuffers==24.3.25
flatland-rl==4.0.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.66.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.11.0
idna==3.7
importlib-resources==1.5.0
ipycanvas==0.13.2
ipyevents==2.0.2
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.8
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pettingzoo==1.24.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.6.0
pyglet==2.0.15
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gitlab==4.4.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ray==2.21.0
recordtype==1.4
redis==5.0.4
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.0
rpds-py==0.18.1
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentry-sdk==2.1.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stable_baselines3==2.3.2
stack-data==0.6.3
SuperSuit==3.9.2
svgutils==0.3.4
swig==4.2.1
sympy==1.13.2
termcolor==2.4.0
threadpoolctl==3.5.0
timeout-decorator==0.5.0
tinyscaler==1.2.7
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.10
wrapt==1.16.0
